Refactor mocks folder
