+++ src/UI/extension/ExtensionLoginButton/tests/helpers/index.tsx
@@ -1,2 +1,1 @@
-export * from './mockExtensionProvider';
 export * from './checkIsLoggedInStore';
