+++ src/UI/extension/ExtensionLoginButton/tests/mocks/index.tsx
@@ -1,0 +1,1 @@
+export * from './mockExtensionProvider';
