+++ src/UI/extension/ExtensionLoginButton/tests/mocks/mockExtensionProvider.ts
